Backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
python-dotenv
pyjwt